requirements.txt
discord.py>=2.3.2
motor>=3.3.2
pymongo>=4.6.0
Pillow>=10.2.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
